# Conflicts:
#	src/game/GameEngine.java
#	src/game/GameLevel.java
#	src/game/sprites/SlowMoIteam.java
#	src/game/sprites/SpeedIteam.java
#	src/uicomponents/game/GameDisplay.java
#	src/uicomponents/game/IteamPane.java
#	src/uicomponents/game/WorldPane.java

+++ src/uicomponents/game/IteamPane.java
@@ -8,7 +8,7 @@ import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.StackPane;
 
 public class IteamPane extends StackPane {
-  public IteamPane( int canWidth, int canHeight ) {
+  public IteamPane( double canWidth, double canHeight ) {
     this.setHeight( canHeight );
     this.setWidth( canWidth );
     this.setStyle( "-fx-background-color:rgba(0, 0, 0, 0);" );
